+++ packages/firebase_messaging/test/firebase_messaging_test.dart
@@ -2,6 +2,8 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
+// ignore_for_file: deprecated_member_use_from_same_package
+
 import 'dart:async';
 
 import 'package:flutter/services.dart';
@@ -100,6 +102,42 @@ void main() {
     iosSettingsFromStream = firebaseMessaging.onIosSettingsRegistered.first;
     await handler(MethodCall('onIosSettingsRegistered', iosSettings.toMap()));
     expect((await iosSettingsFromStream).toMap(), iosSettings.toMap());
+  });
+
+  test('getLaunchMessage', () async {
+    final MethodChannel channel =
+        const MethodChannel('plugins.flutter.io/firebase_messaging');
+    firebaseMessaging =
+        FirebaseMessaging.private(channel, const LocalPlatform());
+
+    channel.setMockMethodCallHandler((MethodCall methodCall) async {
+      switch (methodCall.method) {
+        case 'getLaunchMessage':
+          return <dynamic, dynamic>{
+            'notification': <dynamic, dynamic>{
+              'title': 'Title',
+              'body': 'Body',
+              'click_action': 'FLUTTER_NOTIFICATION_CLICK',
+            },
+            'data': <dynamic, dynamic>{
+              'variable1': 'value1',
+              'variable2': 'value2',
+            },
+          };
+        default:
+          return null;
+      }
+    });
+
+    final Map<dynamic, dynamic> message =
+        await firebaseMessaging.getLaunchMessage();
+
+    expect(message['notification']['title'], 'Title');
+    expect(message['notification']['body'], 'Body');
+    expect(
+        message['notification']['click_action'], 'FLUTTER_NOTIFICATION_CLICK');
+    expect(message['data']['variable1'], 'value1');
+    expect(message['data']['variable2'], 'value2');
   });
 
   test('incoming messages', () async {